+++ src/settings/CustomFieldsSettings.js
@@ -1,8 +1,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import { Switch, Route } from 'react-router-dom';
 
 import { useStripes } from '@folio/stripes/core';
-import { ViewCustomFieldsSettings } from '@folio/stripes/smart-components';
+import { ViewCustomFieldsSettings, EditCustomFieldsSettings } from '@folio/stripes/smart-components';
 
 const propTypes = {
   history: PropTypes.shape({
@@ -16,9 +17,7 @@ const CustomFieldsSettings = ({
 }) => {
   const stripes = useStripes();
 
-  const redirectToEdit = () => {
-    history.push('/users/custom-fields/edit');
-  };
+  const base = '/settings/users/custom-fields';
 
   const permissions = {
     canView: stripes.hasPerm('ui-users.settings.customfields.view'),
@@ -26,17 +25,33 @@ const CustomFieldsSettings = ({
     canDelete: stripes.hasPerm('ui-users.settings.customfields.all'),
   };
 
+  const backendModuleName = 'users';
+  const entityType = 'user';
+
+
   if (!permissions.canView) {
     history.replace('/settings/users');
   }
 
   return (
-    <ViewCustomFieldsSettings
-      backendModuleName="users"
-      entityType="user"
-      redirectToEdit={redirectToEdit}
-      permissions={permissions}
-    />
+    <Switch>
+      <Route exact path={base}>
+        <ViewCustomFieldsSettings
+          backendModuleName={backendModuleName}
+          entityType={entityType}
+          editRoute={`${base}/edit`}
+          permissions={permissions}
+        />
+      </Route>
+      <Route exact path={`${base}/edit`}>
+        <EditCustomFieldsSettings
+          backendModuleName={backendModuleName}
+          entityType={entityType}
+          viewRoute={base}
+          permissions={permissions}
+        />
+      </Route>
+    </Switch>
   );
 };