+++ src/views/UserSearch/Filters.js
@@ -24,6 +24,9 @@ export default class Filters extends React.Component {
     activeFilters: PropTypes.object,
     resources: PropTypes.object.isRequired,
     onChangeHandlers: PropTypes.object.isRequired,
+    resultOffset: PropTypes.shape({
+      replace: PropTypes.func.isRequired,
+    }),
   };
 
   static defaultProps = {
@@ -39,8 +42,7 @@ export default class Filters extends React.Component {
 
   getValuesFromResources = (type, key) => {
     const items = get(this.props.resources, `${type}.records`, [])
-      .map(item => ({ label: item[key], value: item[key] }));
-
+      .map(item => ({ label: item[key], value: item.id }));
     return sortBy(items, 'label');
   };
 
@@ -48,7 +50,12 @@ export default class Filters extends React.Component {
     const {
       activeFilters,
       onChangeHandlers,
+      resultOffset,
     } = this.props;
+
+    if (resultOffset) {
+      resultOffset.replace(0);
+    }
 
     onChangeHandlers.state({
       ...activeFilters,