+++ test/bigtest/interactors/loan-actions-history.js
@@ -6,6 +6,8 @@ import {
   property,
 } from '@bigtest/interactor';
 
+import MultiColumnListInteractor from '@folio/stripes-components/lib/MultiColumnList/tests/interactor'; // eslint-disable-line
+
 import KeyValue from './KeyValue';
 
 @interactor class LoanActionsHistory {
@@ -26,6 +28,7 @@ import KeyValue from './KeyValue';
   isDeclareLostButtonDisabled = property('[data-test-declare-lost-button]', 'disabled');
   claimReturnedButton = scoped('[data-test-claim-returned-button]', ButtonInteractor);
   isClaimReturnedButtonDisabled = property('[data-test-claim-returned-button]', 'disabled');
+  loanActions = scoped('#list-loanactions', MultiColumnListInteractor);
 
   whenLoaded() {
     return this.when(() => this.isPresent).timeout(5000);