+++ test/bigtest/network/serializers/patron-block-limits.js
@@ -1,0 +1,15 @@
+import ApplicationSerializer from './application';
+
+const { isArray } = Array;
+
+export default ApplicationSerializer.extend({
+  serialize(...args) {
+    const json = ApplicationSerializer.prototype.serialize.apply(this, args);
+
+    if (isArray(json.patronBlockLimits)) {
+      return { ...json, totalRecords: json.patronBlockLimits.length };
+    }
+
+    return json.patronBlockLimits;
+  }
+});