+++ test/bigtest/tests/user-edit-page-test.js
@@ -223,7 +223,7 @@ describe('User Edit Page', () => {
 
       it('should display inactive status', () => {
         expect(InstanceViewPage.userInfo.keyValues(5).text).to.equal('inactive');
-      });
+      }).timeout(6000);
     });
 
     describe('clearing expirationDate field', () => {
@@ -236,7 +236,7 @@ describe('User Edit Page', () => {
 
       it('should display active status', () => {
         expect(InstanceViewPage.userInfo.keyValues(5).text).to.equal('active');
-      });
+      }).timeout(6000);
     });
   });
 });