+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [3.1.0] (IN PROGRESS)
 
+* Support `loan-storage` interface version 7.0. UIU-1598.
 * Add posibility to view/create/edit custom fields on user record. UIU-1279.
 * Increase limit of service points in the Associated Service Points dropdown at Settings --> Users --> Fee/Fine Owners. Refs UIU-1540.
 * Fix ability to mark user as inactive. Fixes UIU-913.
@@ -10,6 +11,17 @@
 * Fix bug preventing closed loans list from loading properly when an item has been deleted. Fixes UIU-1411.
 * Fix import path to stripes util. Fixes UIU-1515.
 * Use localized permission names. Refs UIU-488.
+* Create Patron Blocks Limits Table in Settings -> Users. Refs UIU-1167.
+* Add `Load more` button at the end of the result list. Fixes UIU-1532.
+* Increase limit of patron fee/fines, owners and patron groups. Refs UIU-1585.
+* Build filter string by using ids instead of names. Fixes UIU-1596.
+* Add UI to mark claimed returned items as missing. Refs UIU-1216.
+* Fix routing loop caused by incorrect Custom Fields routing. Refs UIU-1594.
+* Settings > Users > Create/Edit Permission set. Refs UIU-1587.
+* Add `Closed loan` translation. Part of UIU-1603.
+* Pass current `servicePointId` to `declare-item-lost`. Part of UIU-1203.
+* Preserve filters after user is edited. Fixes UIU-1604.
+* Remove hardcoded ids from Parton Block Conditions page. Refs UIU-1609.
 
 ## [3.0.0](https://github.com/folio-org/ui-users/tree/v3.0.0) (2020-03-17)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v2.26.0...v3.0.0)